Small changes on Library and the product page
Removed the unneeded info from the Library listing (title, description, price, size)
Removed the space charachter from the end of the components folder, because of these I needed to readd all the files inside, without actual changes.
I limited the related products to 4.
Changed the + link to Line By Line to a text.